chore: Consolidate the refactoring and documentation views into one
This way users won't accidentally have both open with different refactoring states. Also it will make for a cleaner UX since we won't pop up several tabs in unexpected places (ViewColumn)
