+++ src/codescene-tab/webview/refactoring-styles.css
@@ -77,8 +77,3 @@
   margin-top: 12px;
 }
 
-.accept-reject-buttons {
-  display: inline-flex;
-  gap: 10px;
-  margin-top: 12px;
-}
