--- src/documentation/styles.css
@@ -1,7 +1,0 @@
-.documentation-header {
-  margin-bottom: 16px;
-}
-
-#refactoring-button {
-  margin-top: 5px;
-}
